# packages/shared/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/characters": {
      "post": {
        "operationId": "CharacterController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Character created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftCharacterResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new character",
        "tags": [
          "characters"
        ]
      },
      "get": {
        "operationId": "CharacterController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of characters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all characters for the current user",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/drafts/list": {
      "get": {
        "operationId": "CharacterController_findDrafts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of draft characters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DraftCharacterResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all draft (unfinished) characters for the current user",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/created/list": {
      "get": {
        "operationId": "CharacterController_findCreated",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of finished characters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all finished characters for the current user",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/{characterId}": {
      "get": {
        "operationId": "CharacterController_findOne",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Character found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Character not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific character by ID",
        "tags": [
          "characters"
        ]
      },
      "put": {
        "operationId": "CharacterController_update",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "description": "ID of the character to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Fields to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCharacterRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Character updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Character not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a character",
        "tags": [
          "characters"
        ]
      },
      "delete": {
        "operationId": "CharacterController_delete",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Character deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Character not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a character",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/{characterId}/kill": {
      "post": {
        "operationId": "CharacterController_kill",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillCharacterBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Character marked as deceased",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Character not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a character as deceased",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/deceased": {
      "get": {
        "operationId": "CharacterController_getDeceased",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of deceased characters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeceasedCharacterResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all deceased characters",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/character/avatar/generate": {
      "post": {
        "operationId": "CharacterAvatarController_generateAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAvatarRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Avatar generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or avatar generation failed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate character avatar from description",
        "tags": [
          "character"
        ]
      }
    },
...
  },
  "info": {
    "title": "RPG Gen Portal",
    "description": "Prototype API for RPG portal using Gemini tokens",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TacticalStats": {
        "type": "object",
        "properties": {
          "vigor": {
            "type": "number",
            "description": "Vigor stat"
          },
          "finesse": {
            "type": "number",
            "description": "Finesse stat"
          },
          "mind": {
            "type": "number",
            "description": "Mind stat"
          },
          "survival": {
            "type": "number",
            "description": "Survival stat"
          }
        },
        "required": [
          "vigor",
          "finesse",
          "mind",
          "survival"
        ]
      },
...
      "RaceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Race ID (unique identifier)",
            "example": "humain"
          },
          "name": {
            "type": "string",
            "description": "Display name",
            "example": "Humain"
          },
          "trait": {
            "type": "string",
            "description": "Trait name",
            "example": "Polyvalent"
          },
          "traitEffect": {
            "description": "Trait effect details",
            "allOf": [
              {
                "$ref": "#/components/schemas/TraitEffectResponseDto"
              }
            ]
          },
          "bonuses": {
            "description": "Stat bonuses",
            "allOf": [
              {
                "$ref": "#/components/schemas/RaceBonusesResponseDto"
              }
            ]
          },
          "descriptionForAi": {
            "type": "string",
            "description": "Description for AI",
            "example": "Polyvalent, gagne +1 PA au premier tour."
          },
          "color": {
            "type": "string",
            "description": "UI color (hex)",
            "example": "#3b82f6"
          },
          "icon": {
            "type": "string",
            "description": "Icon emoji",
            "example": "👤"
          }
        },
        "required": [
          "id",
          "name",
          "trait",
          "traitEffect",
          "bonuses",
          "color",
          "icon"
        ]
      },
      "SkillResponseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Skill name"
          },
          "proficient": {
            "type": "boolean",
            "description": "Is proficient in this skill"
          },
          "modifier": {
            "type": "number",
            "description": "Skill modifier"
          }
        }
      },
...
      "DraftCharacterResponseDto": {
        "type": "object",
        "properties": {
          "characterId": {
            "type": "string",
            "description": "Unique character ID (UUID)"
          },
          "name": {
            "type": "string",
            "description": "Character name"
          },
          "physicalDescription": {
            "type": "string",
            "description": "Physical description of the character"
          },
          "race": {
            "description": "Character race (new system)",
            "allOf": [
              {
                "$ref": "#/components/schemas/RaceResponseDto"
              }
            ]
          },
          "hp": {
            "type": "number",
            "description": "Current hit points"
          },
          "hpMax": {
            "type": "number",
            "description": "Maximum hit points"
          },
          "totalXp": {
            "type": "number",
            "description": "Total experience points"
          },
          "skills": {
            "description": "Character skills",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillResponseDto"
            }
          },
          "portrait": {
            "type": "string",
            "description": "Character portrait URL or base64"
          },
          "gender": {
            "type": "string",
            "description": "Character gender"
          },
          "inspirationPoints": {
            "type": "number",
            "description": "Inspiration points"
          },
          "isDeceased": {
            "type": "boolean",
            "description": "Whether character is deceased"
          },
          "diedAt": {
            "type": "string",
            "description": "Date of death (ISO string)"
          },
          "deathLocation": {
            "type": "string",
            "description": "Location where character died"
          },
          "state": {
            "type": "string",
            "description": "Character state",
            "enum": [
              "draft",
              "created"
            ]
          },
          "inventory": {
            "description": "Character inventory",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryItemDto"
            }
          },
          "aptitudes": {
            "description": "Character aptitudes (learned abilities)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AptitudeResponseDto"
            }
          },
          "className": {
            "type": "string",
            "description": "Character class (guerrier, rogue, mage)",
            "enum": [
              "guerrier",
              "rogue",
              "mage"
            ]
          },
          "level": {
            "type": "number",
            "description": "Character level (1-20)"
          },
          "raceId": {
            "type": "string",
            "description": "Race ID (humain, nain, elfe, dark_elfe, orc)",
            "enum": [
              "humain",
              "nain",
              "elfe",
              "dark_elfe",
              "orc"
            ]
          },
          "stats": {
            "description": "Tactical stats (vigor, finesse, mind, survival)",
            "allOf": [
              {
                "$ref": "#/components/schemas/TacticalStats"
              }
            ]
          },
          "pa": {
            "type": "number",
            "description": "Current action points"
          },
          "paMax": {
            "type": "number",
            "description": "Maximum action points"
          },
          "pm": {
            "type": "number",
            "description": "Current movement points"
          },
          "pmMax": {
            "type": "number",
            "description": "Maximum movement points"
          },
          "talentPoints": {
            "type": "number",
            "description": "Unspent talent points"
          },
          "voies": {
            "description": "Talent tree progression",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoieProgressDto"
            }
          }
        },
        "required": [
          "characterId",
          "isDeceased",
          "state"
        ]
      },
      "CharacterResponseDto": {
        "type": "object",
        "properties": {
          "characterId": {
            "type": "string",
            "description": "Unique character ID (UUID)"
          },
          "name": {
            "type": "string",
            "description": "Character name"
          },
          "physicalDescription": {
            "type": "string",
            "description": "Physical description of the character"
          },
          "race": {
            "description": "Character race (new system)",
            "allOf": [
              {
                "$ref": "#/components/schemas/RaceResponseDto"
              }
            ]
          },
          "hp": {
            "type": "number",
            "description": "Current hit points"
          },
          "hpMax": {
            "type": "number",
            "description": "Maximum hit points"
          },
          "totalXp": {
            "type": "number",
            "description": "Total experience points"
          },
          "skills": {
            "description": "Character skills",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillResponseDto"
            }
          },
          "portrait": {
            "type": "string",
            "description": "Character portrait URL or base64"
          },
          "gender": {
            "type": "string",
            "description": "Character gender"
          },
          "inspirationPoints": {
            "type": "number",
            "description": "Inspiration points"
          },
          "isDeceased": {
            "type": "boolean",
            "description": "Whether character is deceased"
          },
          "diedAt": {
            "type": "string",
            "description": "Date of death (ISO string)"
          },
          "deathLocation": {
            "type": "string",
            "description": "Location where character died"
          },
          "state": {
            "type": "string",
            "description": "Character state",
            "enum": [
              "draft",
              "created"
            ]
          },
          "inventory": {
            "description": "Character inventory",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryItemDto"
            }
          },
          "aptitudes": {
            "description": "Character aptitudes (learned abilities)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AptitudeResponseDto"
            }
          },
          "className": {
            "type": "string",
            "description": "Character class (guerrier, rogue, mage)",
            "enum": [
              "guerrier",
              "rogue",
              "mage"
            ]
          },
          "level": {
            "type": "number",
            "description": "Character level (1-20)"
          },
          "raceId": {
            "type": "string",
            "description": "Race ID (humain, nain, elfe, dark_elfe, orc)",
            "enum": [
              "humain",
              "nain",
              "elfe",
              "dark_elfe",
              "orc"
            ]
          },
          "stats": {
            "description": "Tactical stats (vigor, finesse, mind, survival)",
            "allOf": [
              {
                "$ref": "#/components/schemas/TacticalStats"
              }
            ]
          },
          "pa": {
            "type": "number",
            "description": "Current action points"
          },
          "paMax": {
            "type": "number",
            "description": "Maximum action points"
          },
          "pm": {
            "type": "number",
            "description": "Current movement points"
          },
          "pmMax": {
            "type": "number",
            "description": "Maximum movement points"
          },
          "talentPoints": {
            "type": "number",
            "description": "Unspent talent points"
          },
          "voies": {
            "description": "Talent tree progression",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoieProgressDto"
            }
          }
        },
        "required": [
          "characterId",
          "isDeceased",
          "state"
        ]
      },
      "UpdateCharacterRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Character name"
          },
          "physicalDescription": {
            "type": "string",
            "description": "Physical description of the character"
          },
          "race": {
            "description": "Character race (new system)",
            "allOf": [
              {
                "$ref": "#/components/schemas/RaceResponseDto"
              }
            ]
          },
          "hp": {
            "type": "number",
            "description": "Current hit points"
          },
          "hpMax": {
            "type": "number",
            "description": "Maximum hit points"
          },
          "totalXp": {
            "type": "number",
            "description": "Total experience points"
          },
          "skills": {
            "description": "Character skills",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillResponseDto"
            }
          },
          "portrait": {
            "type": "string",
            "description": "Character portrait URL or base64"
          },
          "gender": {
            "type": "string",
            "description": "Character gender"
          },
          "inspirationPoints": {
            "type": "number",
            "description": "Inspiration points"
          },
          "state": {
            "type": "string",
            "description": "Character state",
            "enum": [
              "draft",
              "created"
            ]
          },
          "inventory": {
            "description": "Character inventory",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryItemDto"
            }
          },
          "aptitudes": {
            "description": "Character aptitudes (learned abilities)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AptitudeResponseDto"
            }
          },
          "className": {
            "type": "string",
            "description": "Character class (guerrier, rogue, mage)",
            "enum": [
              "guerrier",
              "rogue",
              "mage"
            ]
          },
          "level": {
            "type": "number",
            "description": "Character level (1-20)"
          },
          "raceId": {
            "type": "string",
            "description": "Race ID (humain, nain, elfe, dark_elfe, orc)",
            "enum": [
              "humain",
              "nain",
              "elfe",
              "dark_elfe",
              "orc"
            ]
          },
          "stats": {
            "description": "Tactical stats (vigor, finesse, mind, survival)",
            "allOf": [
              {
                "$ref": "#/components/schemas/TacticalStats"
              }
            ]
          },
          "pa": {
            "type": "number",
            "description": "Current action points"
          },
          "paMax": {
            "type": "number",
            "description": "Maximum action points"
          },
          "pm": {
            "type": "number",
            "description": "Current movement points"
          },
          "pmMax": {
            "type": "number",
            "description": "Maximum movement points"
          },
          "talentPoints": {
            "type": "number",
            "description": "Unspent talent points"
          },
          "voies": {
            "description": "Talent tree progression",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoieProgressDto"
            }
          }
        }
      },
      "KillCharacterBodyDto": {
        "type": "object",
        "properties": {
          "deathLocation": {
            "type": "string",
            "description": "Location where character died"
          }
        }
      },
...
      "GenerateAvatarRequestDto": {
        "type": "object",
        "properties": {
          "characterId": {
            "type": "string",
            "description": "UUID of the character"
          }
        },
        "required": [
          "characterId"
        ]
      },
      "AvatarResponseDto": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string",
            "description": "Generated avatar image URL or base64 data"
          }
        },
        "required": [
          "imageUrl"
        ]
      },
...
    }
  }
}